requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-chartjs==2.3.0
django-extensions==3.1.5
django-import-export==2.8.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
xlrd==2.0.1
xlwt==1.3.0
xhtml2pdf==0.2.7
django-crispy-forms==1.14.0
geopy==2.2.0